:IPy without args acts as --existing (closes #2)
For this to work you have to be running IPython from git, until 0.12
comes out. An error message will be reported for IPython 0.11, since it
did not support --existing.

Those who wish to gain this functionality in IPython 0.11 should take a
look at the unmerged pull request #2:
https://github.com/ivanov/vim-ipython/pull/2

There, Philip Neustrom exploits the fact that on 0.11, IPython's full
connection string could be inferred from it's process listing via ps.
That approach does not work, nor is it necessary, with the connection
file mechanisms now in IPython.