intelligent dedenting, thx @unpingco (closes #15)
before this patch <M-s> in visual mode would just remove the indentation
from all of the selected lines. With this, it removes the indentation
level of the first line in the selection, so that further indentation is
preserved, but this block does not produce indentation errors when sent
to IPython.